update_by_empenho bugfix: creating new execucao for new empenho.subelemento
when filter_by_indexer returns more than one execucao and there's no one
without subelemento. A new execucao should be created with same data as
another one with the same indexer, but with correct subelemento and
empenhado_liquido (from Empenho) and orcado_atualizado == 0 (so it won't
affect the Sum of orcado_total) It won't an execucao with
orcado_atualizado == 0 to appear on the chart, because, at subelemento
level, we show the values of the previous level (elemento), so it will
always be a sum.